+++ src/main/java/com/ssysitel/projectapp/model/DeviceId.java
@@ -16,7 +16,7 @@ public class DeviceId implements Serializable {
     @Column(name = "device_id")
     Long deviceID;
     @Column(name = "user_id")
-    Long userID;
+    Integer userID;
     @Column(name = "account_id")
     String accountId;
